+++ src/main/java/com/listywave/list/repository/list/custom/CustomListRepository.java
@@ -1,5 +1,7 @@
 package com.listywave.list.repository.list.custom;
 
+import com.listywave.collaborator.application.domain.Collaborator;
+import com.listywave.list.application.domain.category.CategoryType;
 import com.listywave.list.application.domain.list.ListEntity;
 import com.listywave.user.application.domain.User;
 import java.util.List;
@@ -13,4 +15,13 @@ public interface CustomListRepository {
     Slice<ListEntity> getRecentLists(Long cursorId, Pageable pageable);
 
     Slice<ListEntity> getRecentListsByFollowing(List<User> followingUsers, Long cursorId, Pageable pageable);
+
+    Slice<ListEntity> findFeedLists(
+            List<Collaborator> collaborators,
+            Long userId,
+            String type,
+            CategoryType category,
+            Long cursorId,
+            Pageable pageable
+    );
 }
